backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
ollama==0.6.1
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
pydantic==2.0.0
